client reader

+++ ClientReader.ps1
@@ -9,6 +9,7 @@ Import-Module /home/pi/PiHomeControl/PiHomeControl.psm1 -Force
 modprobe btusb
 systemctl stop bluetooth
 systemctl start bluetooth
+Get-Job -Name BTScan | Remove-Job -Force
 $ScriptBlock = { hcitool lescan --duplicates > /home/pi/PiHomeControl/BTScan.results }
 Start-Job -name BTScan -ScriptBlock $ScriptBlock
 $BluetoothDevices = @{}
@@ -17,37 +18,40 @@ do
 {
     if (Test-Path /home/pi/PiHomeControl/BTScan.reading) { Remove-Item /home/pi/PiHomeControl/BTScan.reading }
     Start-Sleep -Seconds 20
-    Rename-Item /home/pi/PiHomeControl/BTScan.results BTScan.reading
-    foreach ($Line in (Get-Content /home/pi/PiHomeControl/BTScan.reading))
-    {
-        if ($Line.Split(' ')[0] -match '[0-9a-fA-F][0-9a-fA-F]:[0-9a-fA-F][0-9a-fA-F]:[0-9a-fA-F][0-9a-fA-F]:[0-9a-fA-F][0-9a-fA-F]:[0-9a-fA-F][0-9a-fA-F]:[0-9a-fA-F][0-9a-fA-F]')
+    if (Test-Path /home/pi/PiHomeControl/BTScan.reading) 
+    { 
+        Rename-Item /home/pi/PiHomeControl/BTScan.results BTScan.reading
+        foreach ($Line in (Get-Content /home/pi/PiHomeControl/BTScan.reading))
         {
-            if ($Line.Split(' ')[1] -match 'CC-RT-M-BLE') 
-            { 
-                if ($BluetoothDevices.Keys -notcontains $Line.Split(' ')[0]) 
+            if ($Line.Split(' ')[0] -match '[0-9a-fA-F][0-9a-fA-F]:[0-9a-fA-F][0-9a-fA-F]:[0-9a-fA-F][0-9a-fA-F]:[0-9a-fA-F][0-9a-fA-F]:[0-9a-fA-F][0-9a-fA-F]:[0-9a-fA-F][0-9a-fA-F]')
+            {
+                if ($Line.Split(' ')[1] -match 'CC-RT-M-BLE') 
                 { 
-                    $BluetoothDevices.Add($Line.Split(' ')[0],$Line.Split(' ')[1].Trim()) 
+                    if ($BluetoothDevices.Keys -notcontains $Line.Split(' ')[0]) 
+                    { 
+                        $BluetoothDevices.Add($Line.Split(' ')[0],$Line.Split(' ')[1].Trim()) 
+                    }
+                    else 
+                    {
+                        $BluetoothDevices.($Line.Split(' ')[0]) = $Line.Split(' ')[1]
+                    }
+                    Write-Host "Updating PostreSQL"
+                    $MACAddress = $Line.Split(' ')[0].Trim()
+                    $Statement = "INSERT INTO eq3thermostats (eq3macaddress) SELECT '$MACAddress'";
+                    Write-ToPostgreSQL -Statement $Statement -DBServer $DBServer -DBName $DBName -DBPort 5432 -DBUser $DBUser -DBPassword $DBPassword
+        
                 }
                 else 
                 {
-                    $BluetoothDevices.($Line.Split(' ')[0]) = $Line.Split(' ')[1]
+                    if ($BluetoothDevices.Keys -notcontains $Line.Split(' ')[0]) 
+                    { 
+                        $BluetoothDevices.Add($Line.Split(' ')[0],$Line.Split(' ')[1].Trim()) 
+                    }    
                 }
-                Write-Host "Updating PostreSQL"
-                $MACAddress = $Line.Split(' ')[0].Trim()
-                $Statement = "INSERT INTO eq3thermostats (eq3macaddress) SELECT '$MACAddress'";
-                Write-ToPostgreSQL -Statement $Statement -DBServer $DBServer -DBName $DBName -DBPort 5432 -DBUser $DBUser -DBPassword $DBPassword
-    
-            }
-            else 
-            {
-                if ($BluetoothDevices.Keys -notcontains $Line.Split(' ')[0]) 
-                { 
-                    $BluetoothDevices.Add($Line.Split(' ')[0],$Line.Split(' ')[1].Trim()) 
-                }    
             }
         }
+        Remove-Item /home/pi/PiHomeControl/BTScan.reading
     }
-    Remove-Item /home/pi/PiHomeControl/BTScan.reading
 } until ($Something)
 
 <#foreach ($MACAddress in (Get-EQ3Thermostats))
